pyproject.toml
[project]
name = "async-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "loguru>=0.7.3",
    "pydantic-settings>=2.7.1",
    "sqlalchemy>=2.0.38",
    "aiosqlite>=0.21.0",
]
